+++ resources/views/bootstrap-4/components/buttons/form-button.blade.php
@@ -39,5 +39,5 @@
 @endpush
 
 @if ($confirm !== null)
-    <x-confirm-modal :id="'confirm-modal-'.$confirmId" :title="$confirmTitle" />
+    <x-confirm-modal :id="'confirm-modal-'.$confirmId" :title="$confirmTitle" :confirmVariant="$confirmVariant" />
 @endif