+++ resources/views/bootstrap-5/components/buttons/simple-button.blade.php
@@ -42,5 +42,5 @@
 </button>
 
 @if ($confirm !== null)
-    <x-confirm-modal :id="'confirm-modal-'.$confirmId" :title="$confirmTitle" />
+    <x-confirm-modal :id="'confirm-modal-'.$confirmId" :title="$confirmTitle" :confirmVariant="$confirmVariant" />
 @endif