+++ src/Components/Buttons/Actions/Disabled.php
@@ -15,6 +15,8 @@ class Disabled extends FormButton
 
         $this->variant ??= 'warning';
 
+        $this->confirmVariant ??= 'warning';
+
         $this->text ??= Str::ucfirst(trans('status.disabled'));
 
         $this->formId = 'enable-'.($this->formId ?? Str::random(32));