+++ src/Components/Buttons/Actions/Edit.php
@@ -13,6 +13,8 @@ class Edit extends LinkButton
     {
         $this->variant ??= 'primary';
 
+        $this->confirmVariant ??= 'primary';
+
         $this->text ??= Str::ucfirst(trans('action.edit'));
 
         if ($this->confirm !== null) {